pyproject.toml
[project]
name = "kubewhisper"
version = "0.0.1"
description = "An voice controlled AI agent that manages your Kubernetes cluster"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "websockets>=12.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "kubernetes>=29.0.0",
    "click>=8.1.7",
    "loguru>=0.7.3",
]

[tool.hatch.version]
path = "src/kubewhisper/__about__.py"

[dependency-groups]
dev = [
    "hatch>=1.14.0",
    "pytest-asyncio>=0.25.1",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kubewhisper"]

[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default
lint.select = ["E", "F"]
# Allow unused variables when underscore-prefixed
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Python version compatibility
target-version = "py310"
# Line length matches black's default
line-length = 120

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
